Add files via upload
The Excel sheets describe data analysis to understand the interactions of two factors on a dependent variable. For this particular dataset, the factors used were the innoculum type and the type of digestion. the dependent variable was the nutrient type. The analysis was done for several nutrient types examined, for each case that the p-value was less than 0.05 reflects an insignificant contribution of the factors to the dependent variable.